modifieed version used for mercury run of 6/4/12
Former-commit-id: 7d1c80a3a1bb35d840dfd8e9e26734993c6c6ba0

+++ PT/AMPS/src/general/rnd.cpp
@@ -7,7 +7,8 @@
 
 #include "rnd.h"
 
-int RandomNumberGenerator::rndLastSeed=0;
+
+unsigned long int RandomNumberGenerator::rndLastSeed=0;
 
 void rnd_seed(int seed) {
   int thread;
